nox: Replace with mnox
Retrieved from: git://gitosis.stanford.edu/mnox.git
Commit: c49c6d959ca09629dceff3c6d7979aad1880857d

Includes the following build fixes:
- Update to newer SWIG (works with v1.3.40)
- Add missing cstdio,stdint includes
- Remove direct calls to ethernetaddr constructor
- include sys/stat.h in nox_main (fixes umask build error)
- Remove modules that we don't use and won't compile

+++ nox/src/include/netinet++/vlan.hh
@@ -40,7 +40,7 @@ struct vlan
     uint16_t id() const;
     void     set_id(uint16_t id);
     uint8_t  pcp() const;
-    void     set_pcp(uint8_t pcp);
+    void     set_pcp(uint16_t pcp);
 
     uint16_t proto() const;
 
@@ -67,7 +67,7 @@ vlan::pcp() const
 //-----------------------------------------------------------------------------
 inline
 void
-vlan::set_pcp(uint8_t pcp)
+vlan::set_pcp(uint16_t pcp)
 { tci = (tci & ~htons(PCP_MASK)) |
         htons((((uint16_t)pcp) << PCP_SHIFT) & PCP_MASK); }
 //-----------------------------------------------------------------------------
